# idm/generated/openapi/src/main/java/com/onyxdb/idm/generated/openapi/apis/DomainComponentsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.onyxdb.idm.generated.openapi.apis;

import com.onyxdb.idm.generated.openapi.models.BadRequestResponse;
import com.onyxdb.idm.generated.openapi.models.DomainComponentDTO;
import com.onyxdb.idm.generated.openapi.models.NotFoundResponse;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Domain Components", description = "the Domain Components API")
public interface DomainComponentsApi {

    /**
     * POST /api/v1/domain-components : Create a new domain component
     *
     * @param domainComponentDTO  (required)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "createDomainComponent",
        summary = "Create a new domain component",
        tags = { "Domain Components" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DomainComponentDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/domain-components",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<DomainComponentDTO> createDomainComponent(
        @Parameter(name = "DomainComponentDTO", description = "", required = true) @Valid @RequestBody DomainComponentDTO domainComponentDTO
    );


    /**
     * DELETE /api/v1/domain-components/{domainComponentId} : Delete a domain component by ID
     *
     * @param domainComponentId  (required)
     * @return No Content (status code 204)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "deleteDomainComponent",
        summary = "Delete a domain component by ID",
        tags = { "Domain Components" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No Content"),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/domain-components/{domainComponentId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteDomainComponent(
        @Parameter(name = "domainComponentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("domainComponentId") UUID domainComponentId
    );


    /**
     * GET /api/v1/domain-components : Get all domain components
     *
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "getAllDomainComponents",
        summary = "Get all domain components",
        tags = { "Domain Components" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DomainComponentDTO.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/domain-components",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<DomainComponentDTO>> getAllDomainComponents(
        
    );


    /**
     * GET /api/v1/domain-components/{domainComponentId} : Get a domain component by ID
     *
     * @param domainComponentId  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "getDomainComponentById",
        summary = "Get a domain component by ID",
        tags = { "Domain Components" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DomainComponentDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/domain-components/{domainComponentId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<DomainComponentDTO> getDomainComponentById(
        @Parameter(name = "domainComponentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("domainComponentId") UUID domainComponentId
    );


    /**
     * PUT /api/v1/domain-components/{domainComponentId} : Update a domain component by ID
     *
     * @param domainComponentId  (required)
     * @param domainComponentDTO  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "updateDomainComponent",
        summary = "Update a domain component by ID",
        tags = { "Domain Components" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DomainComponentDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/v1/domain-components/{domainComponentId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<DomainComponentDTO> updateDomainComponent(
        @Parameter(name = "domainComponentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("domainComponentId") UUID domainComponentId,
        @Parameter(name = "DomainComponentDTO", description = "", required = true) @Valid @RequestBody DomainComponentDTO domainComponentDTO
    );

}