# idm/src/main/resources/openapi/scheme.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: IDM - Role Based Access Control
servers:
  - url: http://localhost:8080
    description: Local server
paths:
...
  /api/v1/domain-components:
    get:
      summary: Get all domain components
      operationId: getAllDomainComponents
      tags:
        - Domain Components
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DomainComponentDTO'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
    post:
      summary: Create a new domain component
      operationId: createDomainComponent
      tags:
        - Domain Components
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainComponentDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainComponentDTO'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
  /api/v1/domain-components/{domainComponentId}:
    get:
      summary: Get a domain component by ID
      operationId: getDomainComponentById
      tags:
        - Domain Components
      parameters:
        - name: domainComponentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainComponentDTO'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
    put:
      summary: Update a domain component by ID
      operationId: updateDomainComponent
      tags:
        - Domain Components
      parameters:
        - name: domainComponentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainComponentDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainComponentDTO'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
    delete:
      summary: Delete a domain component by ID
      operationId: deleteDomainComponent
      tags:
        - Domain Components
      parameters:
        - name: domainComponentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
...
components:
  schemas:
...
    DomainComponentDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BadRequestResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    NotFoundResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...